chore: move npm package
