+++ packages/runtime/typescript/package.json
@@ -7,7 +7,7 @@
   "repository": {
     "type": "git",
     "url": "https://github.com/nanofuzz/nanofuzz.git",
-    "directory": "packages/runtime"
+    "directory": "packages/runtime/typescript"
   },
   "main": "build/cjs/index.js",
   "module": "build/esm/index.js",
